add user bookmarks deserialize pseudo numbers to u64

+++ src/api/ugoira.rs
@@ -16,7 +16,7 @@ pub struct UgoiraFrame {
     pub delay: u16,
 }
 
-pub fn fetch_ugoira_meta(client: &ureq::Agent, id: u32) -> Result<UgoiraMeta, ApiError> {
+pub fn fetch_ugoira_meta(client: &ureq::Agent, id: u64) -> Result<UgoiraMeta, ApiError> {
     let url = format!(
         "https://www.pixiv.net/ajax/illust/{}/ugoira_meta?lang=en",
         id
